modified modules to work with short arrays(how I am handling dems)

+++ assignment_2/gtopoCompare.c
@@ -1,4 +1,4 @@
-/*FILENAME: pgmCompare.c
+/*FILENAME: gtopoCompare.c
  *FUNCTIONS:
  *	compareContents
  *	compareHeaders
@@ -6,10 +6,10 @@
 
 #include <stdio.h>
 #include <stdlib.h>
-#include "pgmCompare.h"
+#include "gtopoCompare.h"
 
-//compares two float arrays of length widthlength and checks they are identical
-int compareContents(float* file1, float* file2, int widthlength){
+//compares two short arrays of length widthlength and checks they are identical
+int compareContents(short* file1, short* file2, int widthlength){
 	for(int i=0; i<widthlength; i++){
 		if(file1[i] != file2[i]){
 			return 0;
