+++ assignment_2/gtopoCompare.h
@@ -1,2 +1,2 @@
-int compareContents(float* file1, float* file2, int widthheight);
+int compareContents(short* file1, short* file2, int widthheight);
 int compareHeaders(int* headers1, int* headers2);
