+++ assignment_2/gtopoEmpty.h
@@ -1,1 +1,1 @@
-float* createEmpty(int width, int height);
+short* createEmpty(int width, int height);
