+++ assignment_2/gtopoReduceSize.h
@@ -1,1 +1,1 @@
-float* reduceSize(float* file, int oldwidth, int oldheight, int factor);
+short* reduceSize(short* file, int oldwidth, int oldheight, int factor);
